Fix an embarrassing bug in old postprocessing.
I am going to delete this file. This is only for the record. The "age of
maxium ice thickness" in timing and trimlines figures was actually the
age of maximum ice surface elevation. Fortunately the paper has not yet
been published. I now feel that spending several days on improving data
postprocessing for publication was maybe not entirely worthless...