So 10m hypsogram finally works but I prefer 100m.
Using 10-m bands includes additional details (e.g. one can distinguish the
slightly higher-elevation north slope margin) but I think this is unnecessary.
The new 100-cell (or 10-cell for 10-m bands) contour and hatch pattern is
exceedingly detailed and that renders the figure very heavy on 10-m bands.
There are also unneeded striping instead of a clean contour around 3500 m
elevation. So I stick to 100-m bands on both panels (and in the movie).

Finally done with this! Time for some celebratory ouzo.